Move SystemInfo class to it's own namespace location
This avoids Admin folder namespace conflict and prevents this being discovered only via class mapper
